+++ R/usage.R
@@ -18,6 +18,8 @@ showUsage <- function(appDir = getwd(), appName = NULL, account = NULL, server =
                       usageType = "hours", from = NULL, until = NULL, interval = NULL) {
 
   accountDetails <- accountInfo(account, server)
+  checkShinyappsServer(accountDetails$server)
+
   api <- clientForAccount(accountDetails)
 
   # resolve application
@@ -121,6 +123,8 @@ showMetrics <- function(metricSeries,
 accountUsage <- function(account = NULL, server = NULL, usageType = "hours",
                          from = NULL, until = NULL, interval = NULL) {
   accountDetails <- accountInfo(account, server)
+  checkShinyappsServer(accountDetails$server)
+
   api <- clientForAccount(accountDetails)
 
   # get application usage